Issue #719: Reorg subhub client libs, update tests
